requirements.txt
Pillow
exifread
ffmpeg-python
tqdm
sqlite3
